Change sizes to `size_t`, counts to `uint64_t`.
On 32-bit system, a 32-bit size_t should be used for indexing
into objects, however, we want to use 64-bit counters, because
it is very conceivable that more than 1e32 measurements are
accumulated in a single run.

Fixes: #581

+++ alea/include/alps/alea/core.hpp
@@ -238,7 +238,7 @@ struct reducer_setup
     size_t pos;
 
     /** Total number of instances (thread count/MPI size/etc.) */
-    size_t count;
+    uint64_t count;
 
     /** Reductions will yield valid result on this instance */
     bool have_result;
