+++ alea/src/mean.cpp
@@ -58,7 +58,7 @@ mean_acc<T> &mean_acc<T>::operator=(const mean_acc &other)
 }
 
 template <typename T>
-void mean_acc<T>::add(const computed<T> &source, size_t count)
+void mean_acc<T>::add(const computed<T> &source, uint64_t count)
 {
     internal::check_valid(*this);
     source.add_to(view<T>(store_->data().data(), size()));
